requirements.txt
# Data handling
pandas>=2.0.0
numpy>=1.26.0

# Market data
yfinance>=0.2.40

# Backtesting
vectorbt>=0.26.2

# Visualization
matplotlib>=3.8.0

# Configuration
pyyaml>=6.0

# Jupyter support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0

# Utilities
python-dateutil>=2.9.0
